refactor: class structure & execution logic
