--- pipefuser/modules/parallel_utils/__init__.py
@@ -1,4 +1,0 @@
-from .base_wrapper import BaseModule
-from .models.transformers import PipeFuserTransformer2DWrapper
-
-__all__ = ["BaseModule", "PipeFuserTransformer2DWrapper"]
