--- pipefuser/modules/parallel_utils/models/schedulers/__init__.py
@@ -1,5 +1,0 @@
-from .register import PipeFuserSchedulerWrappers
-
-__all__ = [
-    "PipeFuserSchedulerWrappers"
-]
