--- pipefuser/modules/parallel_utils/models/transformers/__init__.py
@@ -1,9 +1,0 @@
-from .register import PipeFuserTransformerWrappers
-from .base_transformer import PipeFuserTransformerBaseWrapper
-from .transformer_2d import PipeFuserTransformer2DWrapper
-
-__all__ = [
-    "PipeFuserTransformerWrappers"
-    "PipeFuserTransformerBaseWrapper",
-    "PipeFuserTransformer2DWrapper",
-]
