+++ pipefuser/refactor/layers/base_layer.py
@@ -1,0 +1,31 @@
+from abc import abstractmethod, ABCMeta
+from typing import List
+
+import torch
+import torch.nn as nn
+
+from pipefuser.refactor.config.config import ParallelConfig, RuntimeConfig
+from pipefuser.refactor.base_wrapper import PipeFuserBaseWrapper
+
+class PipeFuserLayerBaseWrapper(PipeFuserBaseWrapper, metaclass=ABCMeta):
+    activation_cache: torch.Tensor
+    num_pipeline_patch: int
+    
+    def __init__(
+        self, 
+        module: nn.Module,
+        parallel_config: ParallelConfig,
+        runtime_config: RuntimeConfig,
+    ):
+        super().__init__(
+            module=module,
+            parallel_config=parallel_config, 
+            runtime_config=runtime_config
+        )
+        self.activation_cache = None
+        self.num_pipeline_patch = \
+            self.parallel_config.pp_config.num_pipeline_patch
+
+    @abstractmethod
+    def forward(self, *args, **kwargs):
+        pass
