+++ pipefuser/refactor/parallel.py
@@ -1,0 +1,50 @@
+from typing import Type, Union
+from diffusers.pipelines.pipeline_utils import DiffusionPipeline
+
+from pipefuser.refactor.distributed.parallel_state import (
+    init_distributed_environment,
+    initialize_model_parallel,
+)
+from pipefuser.refactor.config.config import EngineConfig
+from pipefuser.logger import init_logger
+from pipefuser.refactor.pipelines.base_pipeline import PipeFuserPipelineBaseWrapper
+from pipefuser.refactor.pipelines.register import PipeFuserPipelineWrapperRegister
+
+logger = init_logger(__name__)
+
+
+
+class Parallel:
+    def __init__(self, engine_config: EngineConfig):
+        init_distributed_environment()
+        initialize_model_parallel(
+            data_parallel_degree=engine_config.parallel_config.dp_degree,
+            classifier_free_guidance_degree=
+                engine_config.parallel_config.cfg_degree,
+            sequence_parallel_degree=engine_config.parallel_config.sp_degree,
+            tensor_parallel_degree=engine_config.parallel_config.tp_degree,
+            pipefusion_parallel_degree=engine_config.parallel_config.pp_degree,
+        )
+
+        self.engine_config = engine_config
+
+    def __call__(
+        self, 
+        pipe: Union[DiffusionPipeline, Type[DiffusionPipeline]],
+    ) -> Union[
+        PipeFuserPipelineBaseWrapper,
+        Type[PipeFuserPipelineBaseWrapper]
+    ]:
+        if isinstance(pipe, type):
+            pipefuser_pipe_class = \
+                PipeFuserPipelineWrapperRegister.get_class(pipe)
+            return pipefuser_pipe_class
+        elif isinstance(pipe, DiffusionPipeline):
+            pipefuser_pipe_wrapper = \
+                PipeFuserPipelineWrapperRegister.get_class(pipe)
+            return pipefuser_pipe_wrapper(
+                pipeline=pipe, 
+                engine_config=self.engine_config
+            )
+        else:
+            raise ValueError(f"Unsupported type {type(pipe)} for pipe")
