+++ pipefuser/refactor/pipelines/__init__.py
@@ -1,0 +1,7 @@
+from .base_pipeline import PipeFuserPipelineBaseWrapper
+from .pixart_alpha import PipeFuserPixArtAlphaPipeline
+
+__all__ = [
+    "PipeFuserPipelineBaseWrapper",
+    "PipeFuserPixArtAlphaPipeline",
+]
